+++ bin/.gitignore
@@ -1,1 +1,4 @@
 /package1/
+/cadastro_animal/
+/calculo_preco/
+/carteira_vacina/
