+++ src/cadastro_animal/TipoAnimal.java
@@ -1,9 +1,11 @@
-package package1;
+package cadastro_animal;
 
 public enum TipoAnimal {
+	
 	/**
 	 * Classe criada para armazenar os tipos de animais que existem no sistema
 	 */
+	
 	SUINO,
 	BOVINO;
 }
