+++ src/carteira_vacina/CarteiraSuina.java
@@ -1,12 +1,14 @@
-package package1;
+package carteira_vacina;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class CarteiraSuina extends CarteiraVacina {
+	
 /**
  * Classe com as vacinas especificas para a espécie Suína
  */
+	
 	public CarteiraSuina() {
 		super.vacinas = new ArrayList<Vacina>();
 		super.vacinas.add(new Vacina(TipoVacina.COLIBACILOSE, null, null));
