+++ src/carteira_vacina/TipoVacina.java
@@ -1,9 +1,11 @@
-package package1;
+package carteira_vacina;
 
 public enum TipoVacina {
-	/**
-	 * Classe criada para armazenar os tipos de vacina que existem dentre as espécies de animais
-	 */
+	
+/**
+ * Classe criada para armazenar os tipos de vacina que existem dentre as espécies de animais
+ */
+	
 	FEBRE_AFTOSA,
 	CLOSTRIDIOSES,
 	BOTULISMO,
